Compara AIC de modelos não estratificados com modelo estratificado

A função comparar_aic_modelo_estratificado() compara o valor de AIC de um modelo global com estratificação com o valor de AIC de modelos ajustados separadamente para caada uma das regiões.

Usage

comparar_aic_modelo_estratificado(..., nome_modelos)

Arguments

...

recebe os objetos que contém os resultados dos modelos distance sampling ajustados aos dados a partir da função ajustar_modelos_Distance().

nome_modelos

recebe um vetor do tipo caracter contendo os nomes dos modelos.

Value

Retorna uma lista contendo uma tibble contendo o nome dos modelos, graus de liberdade e valores de AIC e uma lista contendo o valor de AIC do modelo global (estratificado) e a soma dos valores de AIC dos modelos ajustados individualmente para cada região.

Examples

if (FALSE) { # \dontrun{
# comparar modelo global (com estratificação) e modelos ajustados para cada
# região
resultado_comparacao <- comparar_aic_modelo_estratificado(
cutias_distance_hn$Cosseno,
cutias_tap_distance_hn,
cutias_par_distance_hn,
cutias_ter_distance_hn,
cutias_anf_distance_hn,
nome_modelos = c(
  "Global",
  "Resex Tapajós-Arapiuns",
  "Parna Serra do Prado",
  "Esec da Terra do Meio",
  "Resex do Riozinho do Anfrísio"
 )
)

resultado_comparacao
} # }