Número total de observações para cada espécie

A função contar_n_obs_sp() gera uma tabela contendo o número total de observações para cada espécie.

Usage

contar_n_obs_sp(dados = monitora_aves_masto_florestal)

Arguments

dados

recebe a tibble gerada pela função filtrar_dados(). Por configuração, carrega a base de dados burtos de aves e medios e grandes mamiferos do Pojeto Monitora Componente Florestal monitora_aves_masto_florestal.

Value

Retorna um objeto do tipo tibble contendo três colunas, nome_sp contendo o nome da espécie, nome_sp_abv que contem o nome da espécie abreviado e n contendo o número de observações realizadas para cada espécie

Author

Vitor N. T. Borges-Junior Luciana A. Fusinatto

Examples

# número de observações para cada espécie - dados brutos
contar_n_obs_sp()
#> # A tibble: 198 × 3
#>    nome_sp                    nome_sp_abv        n
#>    <fct>                      <fct>          <int>
#>  1 aburria_cujubi             abur_cuju        124
#>  2 aburria_cumanensis         abur_cuma         39
#>  3 alouatta_belzebul          alou_belz         60
#>  4 alouatta_caraya            alou_cara          2
#>  5 alouatta_discolor          alou_disc         20
#>  6 alouatta_guariba           alou_guar          2
#>  7 alouatta_guariba_clamitans alou_guar_clam     1
#>  8 alouatta_juara             alou_juar         14
#>  9 alouatta_macconnelli       alou_macc        135
#> 10 alouatta_nigerrima         alou_nige        119
#> # ℹ 188 more rows

# número de observações para cada espécie - dados filtrados
dados_filtrados <- filtrar_dados(
dados = monitora_aves_masto_florestal,
  nome_uc == "resex_tapajos_arapiuns"
)

contar_n_obs_sp(dados_filtrados)
#> # A tibble: 60 × 3
#>    nome_sp               nome_sp_abv     n
#>    <fct>                 <fct>       <int>
#>  1 aburria_cujubi        abur_cuju      30
#>  2 alouatta_nigerrima    alou_nige     108
#>  3 aotus_nigriceps       aotu_nigr       5
#>  4 ateles_chamek         atel_cham       1
#>  5 bradypus_variegatus   brad_vari      13
#>  6 cabassous_unicinctus  caba_unic       2
#>  7 callicebus_hoffmannsi call_hoff      48
#>  8 cebus_unicolor        cebu_unic      72
#>  9 cerdocyon_thous       cerd_thou       2
#> 10 cf_crax_sp            cf_crax_sp      7
#> # ℹ 50 more rows