Número total de observações para cada espécie e em cada Unidade de Conservação

A função contar_n_obs_sp_uc() gera uma tabela contendo o número total de observações para cada espécie e em cada Unidade de Conservação.

Usage

contar_n_obs_sp_uc(dados = monitora_aves_masto_florestal)

Arguments

dados

recebe a tibble gerada pela função filtrar_dados(). Por configuração, carrega a base de dados burtos de aves e médios e grandes mamíferos do Pojeto Monitora Componente Florestal monitora_aves_masto_florestal.

Value

Retorna um objeto do tipo tibble contendo cinco colunas, nome_uc que contem o nome da Unidade de Conservação, nome_uc_abv que contem o nome abreviado da Unidadade de Conservação, nome_sp contendo o nome da espécie, nome_sp_abv que contem o nome abreviado da espécie e n contendo o número de observações realizadas para espécie em cada ano.

Author

Vitor N. T. Borges-Junior Luciana A. Fusinatto

Examples

# número de observações para cada espécie e em cada UC - dados brutos
contar_n_obs_sp_uc()
#> # A tibble: 1,103 × 5
#>    nome_uc               nome_uc_abv          nome_sp          nome_sp_abv     n
#>    <fct>                 <fct>                <fct>            <fct>       <int>
#>  1 esec_da_terra_do_meio esec_da_terr_do_meio aburria_cujubi   abur_cuju      10
#>  2 esec_da_terra_do_meio esec_da_terr_do_meio alouatta_discol… alou_disc       1
#>  3 esec_da_terra_do_meio esec_da_terr_do_meio ateles_marginat… atel_marg      14
#>  4 esec_da_terra_do_meio esec_da_terr_do_meio callicebus_viei… call_viei      92
#>  5 esec_da_terra_do_meio esec_da_terr_do_meio crax_fasciolata  crax_fasc      40
#>  6 esec_da_terra_do_meio esec_da_terr_do_meio crypturellus_sp  cryp_sp        78
#>  7 esec_da_terra_do_meio esec_da_terr_do_meio dasyprocta_croc… dasy_croc     334
#>  8 esec_da_terra_do_meio esec_da_terr_do_meio didelphis_marsu… dide_mars       1
#>  9 esec_da_terra_do_meio esec_da_terr_do_meio guerlinguetus_a… guer_aest      89
#> 10 esec_da_terra_do_meio esec_da_terr_do_meio mazama_americana maza_amer      24
#> # ℹ 1,093 more rows

# número de observações para cada espécie e em cada UC - dados filtrados
dados_filtrados <- filtrar_dados(
dados = monitora_aves_masto_florestal,
  nome_uc == "resex_tapajos_arapiuns"
)

contar_n_obs_sp_uc(dados_filtrados)
#> # A tibble: 60 × 5
#>    nome_uc                nome_uc_abv nome_sp               nome_sp_abv     n
#>    <fct>                  <fct>       <fct>                 <fct>       <int>
#>  1 resex_tapajos_arapiuns rese_tapa   aburria_cujubi        abur_cuju      30
#>  2 resex_tapajos_arapiuns rese_tapa   alouatta_nigerrima    alou_nige     108
#>  3 resex_tapajos_arapiuns rese_tapa   aotus_nigriceps       aotu_nigr       5
#>  4 resex_tapajos_arapiuns rese_tapa   ateles_chamek         atel_cham       1
#>  5 resex_tapajos_arapiuns rese_tapa   bradypus_variegatus   brad_vari      13
#>  6 resex_tapajos_arapiuns rese_tapa   cabassous_unicinctus  caba_unic       2
#>  7 resex_tapajos_arapiuns rese_tapa   callicebus_hoffmannsi call_hoff      48
#>  8 resex_tapajos_arapiuns rese_tapa   cebus_unicolor        cebu_unic      72
#>  9 resex_tapajos_arapiuns rese_tapa   cerdocyon_thous       cerd_thou       2
#> 10 resex_tapajos_arapiuns rese_tapa   cf_crax_sp            cf_crax_sp      7
#> # ℹ 50 more rows