Número total de observações em cada Unidade de Conservação

A função contar_n_obs_uc() gera uma tabela contendo o número total de observações para cada Unidade de Conservação.

Usage

contar_n_obs_uc(dados = monitora_aves_masto_florestal)

Arguments

dados

recebe a tibble gerada pela funções filtrar_dados(). Por configuração, carrega a base de dados burtos de aves e médios e grandes mamíferos do Pojeto Monitora Componente Florestal monitora_aves_masto_florestal.

Value

Retorna um objeto do tipo tibble contendo três colunas, nome_uc contendo o nome da Unidade de Conservação, nome_uc_abv que contem o nome abreviado da Unidade de Conservação e n contendo o número de observações realizadas para cada Unidade de Conservação.

Author

Vitor N. T. Borges-Junior Luciana A. Fusinatto

Examples

# número de observações para cada UC - dados brutos
contar_n_obs_uc()
#> # A tibble: 40 × 3
#>    nome_uc                       nome_uc_abv                n
#>    <fct>                         <fct>                  <int>
#>  1 esec_da_terra_do_meio         esec_da_terr_do_meio    1229
#>  2 esec_maraca                   esec_mara                434
#>  3 esec_niquia                   esec_niqu                132
#>  4 esec_serra_geral_do_tocantins esec_serr_gera_do_toca    12
#>  5 flona_do_jamari               flon_do_jama            1044
#>  6 parna_chapada_dos_veadeiros   parn_chap_dos_vead        11
#>  7 parna_da_amazonia             parn_da_amaz            1081
#>  8 parna_da_serra_da_bocaina     parn_da_serr_da_boca      50
#>  9 parna_da_serra_da_bodoquena   parn_da_serr_da_bodo     196
#> 10 parna_da_serra_da_cutia       parn_da_serr_da_cuti     829
#> # ℹ 30 more rows

# número de observações para cada UC - dados filtrados
dados_filtrados <- filtrar_dados(
dados = monitora_aves_masto_florestal,
  nome_uc == "resex_tapajos_arapiuns"
)

contar_n_obs_uc(dados_filtrados)
#> # A tibble: 1 × 3
#>   nome_uc                nome_uc_abv     n
#>   <fct>                  <fct>       <int>
#> 1 resex_tapajos_arapiuns rese_tapa    3582