Número de Unidades de Conservação amostradas por ano

A função contar_n_uc_ano() gera o número total de Unidades de Conservação amostradas por ano contidos na base de dados fornecida.

Usage

contar_n_uc_ano(dados = monitora_aves_masto_florestal)

Arguments

dados

recebe a tibble gerada pela função filtrar_dados(). Por configuraçãoo, carrega a base de dados burtos de aves e médios e grandes mamíferos do Pojeto Monitora Componente Florestal monitora_aves_masto_florestal.

Value

Retorna um objeto do tipo tibble contendo o número total de Unidades de Conservação amostradas por ano contido na base de dados fornecida.

Author

Vitor N. T. Borges-Junior Luciana A. Fusinatto

Examples

# número total de UCs em cada ano - dados brutos
contar_n_uc_ano()
#> # A tibble: 8 × 2
#>     ano n_ucs
#>   <dbl> <int>
#> 1  2014    13
#> 2  2015    13
#> 3  2016    21
#> 4  2017    25
#> 5  2018    28
#> 6  2019    31
#> 7  2020    18
#> 8  2021    24

# número total de UCs em cada ano - dados filtrados
dados_filtrados <- filtrar_dados(
dados = monitora_aves_masto_florestal,
  nome_uc == "resex_tapajos_arapiuns"
)

contar_n_uc_ano(dados_filtrados)
#> # A tibble: 8 × 2
#>     ano n_ucs
#>   <dbl> <int>
#> 1  2014     1
#> 2  2015     1
#> 3  2016     1
#> 4  2017     1
#> 5  2018     1
#> 6  2019     1
#> 7  2020     1
#> 8  2021     1