Agrupas as distâncias em intervalos de tamanhos homogêneos ou diferentes

A função definir_intervalos_distancia() agrupa as distâncias perpendiculares nas quais os indivíduos foram observados em intervalos regulares (ex. mesmo tamanho), ou irregulares (ex. tamanhos diferentes).

Usage

definir_intervalos_distancia(
         dados,
         intervalos_distancia
       )

Arguments

dados

recebe uma tibble gerada pela função transformar_dados_formato_Distance().

intervalos_distancia

recebe um vetor numérico contendo os intervalos a partir dos quais as distâncias serão agrupadas.

Value

Retorna a tibble de entrada contendo duas novas colunas, distbegin

e distend que definem o intervalo de agrupamento das distâncias perpendiculares

Examples

# carregar pacote
library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

# gerar dados filtrados
dasy_croc_tap_arap_repeticao <- filtrar_dados(
dados = monitora_aves_masto_florestal,
  nome_uc == "resex_tapajos_arapiuns",
  nome_sp == "dasyprocta_croconota",
  validacao_obs = "especie"
) |>
transformar_dados_formato_Distance(
  Region.Label,
  Sample.Label,
)

# definir intervalos de distância (binagem)
dasy_croc_tap_arap_repeticao_binados <- dasy_croc_tap_arap_repeticao |>
  definir_intervalos_distancia(intervalos_distancia = seq(
    from = 0,
    to  = 50,
    by = 1.5
  ))
#> Warning: Some distances were outside bins and have been removed.

glimpse(dasy_croc_tap_arap_repeticao_binados)
#> Rows: 1,337
#> Columns: 16
#> $ Region.Label    <fct> resex_tapajos_arapiuns, resex_tapajos_arapiuns, resex_…
#> $ Area            <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ Sample.Label    <fct> boim, boim, boim, boim, boim, boim, boim, boim, boim, …
#> $ repeated_visits <int> 150, 150, 150, 150, 150, 150, 150, 150, 150, 150, 150,…
#> $ Effort_day      <dbl> 5000, 5000, 5000, 5000, 5000, 5000, 5000, 5000, 5000, …
#> $ Effort          <dbl> 750000, 750000, 750000, 750000, 750000, 750000, 750000…
#> $ sampling_day    <dttm> 2015-02-06, 2015-06-28, 2015-06-29, 2015-06-30, 2015-…
#> $ distance        <dbl> 10.00, 20.00, 28.00, 30.00, 4.00, 10.00, 20.00, 6.00, …
#> $ season          <fct> chuvosa, seca, seca, seca, seca, seca, seca, seca, sec…
#> $ year            <dbl> 2015, 2015, 2015, 2015, 2015, 2015, 2015, 2015, 2015, …
#> $ size            <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 2, 1, 1, 1, 1, 1, 1, 1, 1, 3, …
#> $ cense_time      <drtn> 525 mins, 600 mins, 600 mins, 645 mins, 215 mins, 215…
#> $ speed           <dbl> 0.5714286, 0.5000000, 0.5000000, 0.4651163, 1.3953488,…
#> $ object          <int> 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16,…
#> $ distbegin       <dbl> 9.0, 19.5, 27.0, 28.5, 3.0, 9.0, 19.5, 4.5, 9.0, 4.5, …
#> $ distend         <dbl> 10.5, 21.0, 28.5, 30.0, 4.5, 10.5, 21.0, 6.0, 10.5, 6.…

# definir intervalos de distância diferentes

dasy_croc_tap_arap_repeticao_binados2 <- dasy_croc_tap_arap_repeticao |>
definir_intervalos_distancia(
    intervalos_distancia = c(0, seq(
      from = 1,
      to = 55,
      by = 1.4
    ))
  )

glimpse(dasy_croc_tap_arap_repeticao_binados2)
#> Rows: 1,338
#> Columns: 16
#> $ Region.Label    <fct> resex_tapajos_arapiuns, resex_tapajos_arapiuns, resex_…
#> $ Area            <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ Sample.Label    <fct> boim, boim, boim, boim, boim, boim, boim, boim, boim, …
#> $ repeated_visits <int> 150, 150, 150, 150, 150, 150, 150, 150, 150, 150, 150,…
#> $ Effort_day      <dbl> 5000, 5000, 5000, 5000, 5000, 5000, 5000, 5000, 5000, …
#> $ Effort          <dbl> 750000, 750000, 750000, 750000, 750000, 750000, 750000…
#> $ sampling_day    <dttm> 2015-02-06, 2015-06-28, 2015-06-29, 2015-06-30, 2015-…
#> $ distance        <dbl> 10.00, 20.00, 28.00, 30.00, 4.00, 10.00, 20.00, 6.00, …
#> $ season          <fct> chuvosa, seca, seca, seca, seca, seca, seca, seca, sec…
#> $ year            <dbl> 2015, 2015, 2015, 2015, 2015, 2015, 2015, 2015, 2015, …
#> $ size            <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 2, 1, 1, 1, 1, 1, 1, 1, 1, 3, …
#> $ cense_time      <drtn> 525 mins, 600 mins, 600 mins, 645 mins, 215 mins, 215…
#> $ speed           <dbl> 0.5714286, 0.5000000, 0.5000000, 0.4651163, 1.3953488,…
#> $ object          <int> 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16,…
#> $ distbegin       <dbl> 9.4, 19.2, 27.6, 29.0, 3.8, 9.4, 19.2, 5.2, 9.4, 5.2, …
#> $ distend         <dbl> 10.8, 20.6, 29.0, 30.4, 5.2, 10.8, 20.6, 6.6, 10.8, 6.…