Filtra os dados a partir das Unidades de Conservação, das espécies e do nível taxonomico de validação

A função filtrar_dados() filtra a base de dados brutos a partir das colunas nome_uc, nome_sp e validacao que correspondem a(s) Unidade(s) de Conservação, a(s) espécie(s) e ao(s) nível(is) taxonômico(s) de validação desejadas, respectivamente.

Usage

filtrar_dados(
         dados = monitora_aves_masto_florestal,
         nome_ucs = NULL,
         nome_sps = NULL,
         validacao_obs = c("ordem", "familia", "genero", "especie", "na")
       )

Arguments

dados

recebe uma tibble que contenha as colunas nome_uc, nome_sp e validacao. Por configuração, carrega a base de dados burto de aves e médios e grandes mamíferos do Pojeto Monitora Componente Florestal, monitora_aves_mamiferos_florestal.

nome_ucs

recebe um vetor do tipo caracter contendo o nome de uma ou mais Unidades de Conservação.

nome_sps

recebe um vetor do tipo caracter contendo o nome de uma ou mais espécies.

validacao_obs

recebe um vetor do tipo caracter contendo o nome de um ou mais níveis taxonômicos de validação.

Value

Retorna um objeto do tipo tibble contendo os dados do Projeto Monitora Componente Florestal, filtrados a partir da(s) Unidade(s) de Conservação, da(s) espécie(s), ou dos nível(is) taxonômico(s) de validação.

Details

A filtragem da base de dados pode ser feita a partir do nome das Unidades de Conservação, a partir do nome das espécies, a partir do nível taxonômico de validação, ou a partir de suas combinações. Para maiores detalhes consulte os exemplos.

Author

Vitor N. T. Borges-Junior Luciana A. Fusinatto

Examples

# carregar pacote
library(dplyr)

# carregar dados filtrados por uma Unidade de Conservacão
dados_filtrados_uc1 <- filtrar_dados(
nome_ucs = "resex_tapajos_arapiuns"
)

glimpse(dados_filtrados_uc1)
#> Rows: 3,576
#> Columns: 24
#> $ nome_ea             <fct> ana, ana, ana, ana, ana, ana, ana, ana, ana, ana, …
#> $ n_visitas_repetidas <int> 67, 67, 67, 67, 67, 67, 67, 67, 67, 67, 67, 67, 67…
#> $ codigo_uc           <dbl> 259, 259, 259, 259, 259, 259, 259, 259, 259, 259, …
#> $ categoria_uc        <fct> resex, resex, resex, resex, resex, resex, resex, r…
#> $ nome_uc             <fct> resex_tapajos_arapiuns, resex_tapajos_arapiuns, re…
#> $ nome_uc_abv         <fct> dasy_croc, maza_sp, dasy_croc, pene_sp, guer_aest,…
#> $ numero_ea           <dbl> 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6,…
#> $ data_amostragem     <dttm> 2015-02-25, 2015-02-25, 2015-02-25, 2015-02-26, 2…
#> $ estacao             <fct> chuvosa, chuvosa, chuvosa, chuvosa, chuvosa, chuvo…
#> $ ano                 <dbl> 2015, 2015, 2015, 2015, 2015, 2015, 2015, 2015, 20…
#> $ esforco_dia         <dbl> 5000, 5000, 5000, 5000, 5000, 5000, 5000, 5000, 50…
#> $ esforco_total       <dbl> 335000, 335000, 335000, 335000, 335000, 335000, 33…
#> $ nome_classe         <fct> mammalia, mammalia, mammalia, aves, mammalia, mamm…
#> $ nome_ordem          <fct> rodentia, artiodactyla, rodentia, galliformes, rod…
#> $ nome_familia        <fct> dasyproctidae, cervidae, dasyproctidae, cracidae, …
#> $ nome_genero         <fct> dasyprocta, mazama, dasyprocta, penelope, guerling…
#> $ nome_sp             <fct> dasyprocta_croconota, mazama_sp, dasyprocta_crocon…
#> $ nome_sp_abv         <fct> dasy_croc, maza_sp, dasy_croc, pene_sp, guer_aest,…
#> $ validacao           <fct> especie, genero, especie, genero, especie, especie…
#> $ distancia           <dbl> 10, 5, 0, 3, 3, 5, 12, 18, 30, 15, 5, 10, 30, 0, 1…
#> $ tamanho_grupo       <dbl> 1, 1, 1, 2, 1, 1, 1, 1, 20, 2, 5, 1, 5, 22, 10, 2,…
#> $ velocidade_km_h     <dbl> 0.5555556, NA, NA, 0.5555556, 0.5555556, NA, 0.769…
#> $ tempo_censo         <drtn> 540 mins, 540 mins, 540 mins, 540 mins, 540 mins,…
#> $ numero_observadores <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 2, 2,…

# carregar dados filtrados por mais de uma Unidade de Conservação
dados_filtrados_uc2 <- filtrar_dados(
nome_ucs = c("resex_tapajos_arapiuns", "resex_barreiro_das_antas")
)

glimpse(dados_filtrados_uc2)
#> Rows: 4,136
#> Columns: 24
#> $ nome_ea             <fct> ana, ana, ana, ana, ana, ana, ana, ana, ana, ana, …
#> $ n_visitas_repetidas <int> 67, 67, 67, 67, 67, 67, 67, 67, 67, 67, 67, 67, 67…
#> $ codigo_uc           <dbl> 259, 259, 259, 259, 259, 259, 259, 259, 259, 259, …
#> $ categoria_uc        <fct> resex, resex, resex, resex, resex, resex, resex, r…
#> $ nome_uc             <fct> resex_tapajos_arapiuns, resex_tapajos_arapiuns, re…
#> $ nome_uc_abv         <fct> dasy_croc, maza_sp, dasy_croc, pene_sp, guer_aest,…
#> $ numero_ea           <dbl> 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6,…
#> $ data_amostragem     <dttm> 2015-02-25, 2015-02-25, 2015-02-25, 2015-02-26, 2…
#> $ estacao             <fct> chuvosa, chuvosa, chuvosa, chuvosa, chuvosa, chuvo…
#> $ ano                 <dbl> 2015, 2015, 2015, 2015, 2015, 2015, 2015, 2015, 20…
#> $ esforco_dia         <dbl> 5000, 5000, 5000, 5000, 5000, 5000, 5000, 5000, 50…
#> $ esforco_total       <dbl> 335000, 335000, 335000, 335000, 335000, 335000, 33…
#> $ nome_classe         <fct> mammalia, mammalia, mammalia, aves, mammalia, mamm…
#> $ nome_ordem          <fct> rodentia, artiodactyla, rodentia, galliformes, rod…
#> $ nome_familia        <fct> dasyproctidae, cervidae, dasyproctidae, cracidae, …
#> $ nome_genero         <fct> dasyprocta, mazama, dasyprocta, penelope, guerling…
#> $ nome_sp             <fct> dasyprocta_croconota, mazama_sp, dasyprocta_crocon…
#> $ nome_sp_abv         <fct> dasy_croc, maza_sp, dasy_croc, pene_sp, guer_aest,…
#> $ validacao           <fct> especie, genero, especie, genero, especie, especie…
#> $ distancia           <dbl> 10, 5, 0, 3, 3, 5, 12, 18, 30, 15, 5, 10, 30, 0, 1…
#> $ tamanho_grupo       <dbl> 1, 1, 1, 2, 1, 1, 1, 1, 20, 2, 5, 1, 5, 22, 10, 2,…
#> $ velocidade_km_h     <dbl> 0.5555556, NA, NA, 0.5555556, 0.5555556, NA, 0.769…
#> $ tempo_censo         <drtn> 540 mins, 540 mins, 540 mins, 540 mins, 540 mins,…
#> $ numero_observadores <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 2, 2,…

# carregar dados filtrados por mais de uma Unidade de Conservação e níveis taxonômicos de validação
dados_filtrados_uc_validacao1 <- filtrar_dados(
nome_ucs = "resex_tapajos_arapiuns",
validacao_obs = c("genero", "especie")
)

glimpse(dados_filtrados_uc_validacao1)
#> Rows: 3,534
#> Columns: 24
#> $ nome_ea             <fct> ana, ana, ana, ana, ana, ana, ana, ana, ana, ana, …
#> $ n_visitas_repetidas <int> 67, 67, 67, 67, 67, 67, 67, 67, 67, 67, 67, 67, 67…
#> $ codigo_uc           <dbl> 259, 259, 259, 259, 259, 259, 259, 259, 259, 259, …
#> $ categoria_uc        <fct> resex, resex, resex, resex, resex, resex, resex, r…
#> $ nome_uc             <fct> resex_tapajos_arapiuns, resex_tapajos_arapiuns, re…
#> $ nome_uc_abv         <fct> dasy_croc, maza_sp, dasy_croc, pene_sp, guer_aest,…
#> $ numero_ea           <dbl> 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6,…
#> $ data_amostragem     <dttm> 2015-02-25, 2015-02-25, 2015-02-25, 2015-02-26, 2…
#> $ estacao             <fct> chuvosa, chuvosa, chuvosa, chuvosa, chuvosa, chuvo…
#> $ ano                 <dbl> 2015, 2015, 2015, 2015, 2015, 2015, 2015, 2015, 20…
#> $ esforco_dia         <dbl> 5000, 5000, 5000, 5000, 5000, 5000, 5000, 5000, 50…
#> $ esforco_total       <dbl> 335000, 335000, 335000, 335000, 335000, 335000, 33…
#> $ nome_classe         <fct> mammalia, mammalia, mammalia, aves, mammalia, mamm…
#> $ nome_ordem          <fct> rodentia, artiodactyla, rodentia, galliformes, rod…
#> $ nome_familia        <fct> dasyproctidae, cervidae, dasyproctidae, cracidae, …
#> $ nome_genero         <fct> dasyprocta, mazama, dasyprocta, penelope, guerling…
#> $ nome_sp             <fct> dasyprocta_croconota, mazama_sp, dasyprocta_crocon…
#> $ nome_sp_abv         <fct> dasy_croc, maza_sp, dasy_croc, pene_sp, guer_aest,…
#> $ validacao           <fct> especie, genero, especie, genero, especie, especie…
#> $ distancia           <dbl> 10, 5, 0, 3, 3, 5, 12, 18, 30, 15, 5, 10, 30, 0, 1…
#> $ tamanho_grupo       <dbl> 1, 1, 1, 2, 1, 1, 1, 1, 20, 2, 5, 1, 5, 22, 10, 2,…
#> $ velocidade_km_h     <dbl> 0.5555556, NA, NA, 0.5555556, 0.5555556, NA, 0.769…
#> $ tempo_censo         <drtn> 540 mins, 540 mins, 540 mins, 540 mins, 540 mins,…
#> $ numero_observadores <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 2, 2,…

# carregar dados filtrados por uma espécie
dados_filtrados_sp1 <- filtrar_dados(
nome_sps = "dasyprocta_croconota"
)

glimpse(dados_filtrados_sp1)
#> Rows: 2,069
#> Columns: 24
#> $ nome_ea             <fct> ana, ana, ana, ana, ana, ana, ana, ana, ana, ana, …
#> $ n_visitas_repetidas <int> 67, 67, 67, 67, 67, 67, 67, 67, 67, 67, 67, 67, 67…
#> $ codigo_uc           <dbl> 259, 259, 259, 259, 259, 259, 259, 259, 259, 259, …
#> $ categoria_uc        <fct> resex, resex, resex, resex, resex, resex, resex, r…
#> $ nome_uc             <fct> resex_tapajos_arapiuns, resex_tapajos_arapiuns, re…
#> $ nome_uc_abv         <fct> dasy_croc, dasy_croc, dasy_croc, dasy_croc, dasy_c…
#> $ numero_ea           <dbl> 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6,…
#> $ data_amostragem     <dttm> 2015-02-25, 2015-02-25, 2015-02-27, 2015-05-31, 2…
#> $ estacao             <fct> chuvosa, chuvosa, chuvosa, chuvosa, chuvosa, seca,…
#> $ ano                 <dbl> 2015, 2015, 2015, 2015, 2015, 2015, 2015, 2015, 20…
#> $ esforco_dia         <dbl> 5000, 5000, 5000, 5000, 5000, 5000, 5000, 5000, 50…
#> $ esforco_total       <dbl> 335000, 335000, 335000, 335000, 335000, 335000, 33…
#> $ nome_classe         <fct> mammalia, mammalia, mammalia, mammalia, mammalia, …
#> $ nome_ordem          <fct> rodentia, rodentia, rodentia, rodentia, rodentia, …
#> $ nome_familia        <fct> dasyproctidae, dasyproctidae, dasyproctidae, dasyp…
#> $ nome_genero         <fct> dasyprocta, dasyprocta, dasyprocta, dasyprocta, da…
#> $ nome_sp             <fct> dasyprocta_croconota, dasyprocta_croconota, dasypr…
#> $ nome_sp_abv         <fct> dasy_croc, dasy_croc, dasy_croc, dasy_croc, dasy_c…
#> $ validacao           <fct> especie, especie, especie, especie, especie, espec…
#> $ distancia           <dbl> 10, 0, 5, 12, 18, 15, 10, 10, 20, 20, 10, 0, 0, 10…
#> $ tamanho_grupo       <dbl> 1, 1, 1, 1, 1, 2, 1, 2, 1, 2, 1, 1, 1, 1, 2, 1, 2,…
#> $ velocidade_km_h     <dbl> 0.5555556, NA, NA, 0.7692308, NA, NA, NA, 0.625000…
#> $ tempo_censo         <drtn> 540 mins, 540 mins, 540 mins, 390 mins, 390 mins,…
#> $ numero_observadores <dbl> 1, 1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2,…

# carregar dados filtrados por mais de uma especie
dados_filtrados_sp2 <- filtrar_dados(
nome_sps = c("dasyprocta_croconota", "dasyprocta_iacki")
)

glimpse(dados_filtrados_sp2)
#> Rows: 2,075
#> Columns: 24
#> $ nome_ea             <fct> ana, ana, ana, ana, ana, ana, ana, ana, ana, ana, …
#> $ n_visitas_repetidas <int> 67, 67, 67, 67, 67, 67, 67, 67, 67, 67, 67, 67, 67…
#> $ codigo_uc           <dbl> 259, 259, 259, 259, 259, 259, 259, 259, 259, 259, …
#> $ categoria_uc        <fct> resex, resex, resex, resex, resex, resex, resex, r…
#> $ nome_uc             <fct> resex_tapajos_arapiuns, resex_tapajos_arapiuns, re…
#> $ nome_uc_abv         <fct> dasy_croc, dasy_croc, dasy_croc, dasy_croc, dasy_c…
#> $ numero_ea           <dbl> 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6,…
#> $ data_amostragem     <dttm> 2015-02-25, 2015-02-25, 2015-02-27, 2015-05-31, 2…
#> $ estacao             <fct> chuvosa, chuvosa, chuvosa, chuvosa, chuvosa, seca,…
#> $ ano                 <dbl> 2015, 2015, 2015, 2015, 2015, 2015, 2015, 2015, 20…
#> $ esforco_dia         <dbl> 5000, 5000, 5000, 5000, 5000, 5000, 5000, 5000, 50…
#> $ esforco_total       <dbl> 335000, 335000, 335000, 335000, 335000, 335000, 33…
#> $ nome_classe         <fct> mammalia, mammalia, mammalia, mammalia, mammalia, …
#> $ nome_ordem          <fct> rodentia, rodentia, rodentia, rodentia, rodentia, …
#> $ nome_familia        <fct> dasyproctidae, dasyproctidae, dasyproctidae, dasyp…
#> $ nome_genero         <fct> dasyprocta, dasyprocta, dasyprocta, dasyprocta, da…
#> $ nome_sp             <fct> dasyprocta_croconota, dasyprocta_croconota, dasypr…
#> $ nome_sp_abv         <fct> dasy_croc, dasy_croc, dasy_croc, dasy_croc, dasy_c…
#> $ validacao           <fct> especie, especie, especie, especie, especie, espec…
#> $ distancia           <dbl> 10, 0, 5, 12, 18, 15, 10, 10, 20, 20, 10, 0, 0, 10…
#> $ tamanho_grupo       <dbl> 1, 1, 1, 1, 1, 2, 1, 2, 1, 2, 1, 1, 1, 1, 2, 1, 2,…
#> $ velocidade_km_h     <dbl> 0.5555556, NA, NA, 0.7692308, NA, NA, NA, 0.625000…
#> $ tempo_censo         <drtn> 540 mins, 540 mins, 540 mins, 390 mins, 390 mins,…
#> $ numero_observadores <dbl> 1, 1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2,…

# carregar dados filtrados por unidade de conservacao e especie
dados_filtrados_uc_sp <- filtrar_dados(
nome_ucs = "resex_tapajos_arapiuns",
nome_sps = c("dasyprocta_croconota")
)

glimpse(dados_filtrados_uc_sp)
#> Rows: 1,352
#> Columns: 24
#> $ nome_ea             <fct> ana, ana, ana, ana, ana, ana, ana, ana, ana, ana, …
#> $ n_visitas_repetidas <int> 67, 67, 67, 67, 67, 67, 67, 67, 67, 67, 67, 67, 67…
#> $ codigo_uc           <dbl> 259, 259, 259, 259, 259, 259, 259, 259, 259, 259, …
#> $ categoria_uc        <fct> resex, resex, resex, resex, resex, resex, resex, r…
#> $ nome_uc             <fct> resex_tapajos_arapiuns, resex_tapajos_arapiuns, re…
#> $ nome_uc_abv         <fct> dasy_croc, dasy_croc, dasy_croc, dasy_croc, dasy_c…
#> $ numero_ea           <dbl> 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6,…
#> $ data_amostragem     <dttm> 2015-02-25, 2015-02-25, 2015-02-27, 2015-05-31, 2…
#> $ estacao             <fct> chuvosa, chuvosa, chuvosa, chuvosa, chuvosa, seca,…
#> $ ano                 <dbl> 2015, 2015, 2015, 2015, 2015, 2015, 2015, 2015, 20…
#> $ esforco_dia         <dbl> 5000, 5000, 5000, 5000, 5000, 5000, 5000, 5000, 50…
#> $ esforco_total       <dbl> 335000, 335000, 335000, 335000, 335000, 335000, 33…
#> $ nome_classe         <fct> mammalia, mammalia, mammalia, mammalia, mammalia, …
#> $ nome_ordem          <fct> rodentia, rodentia, rodentia, rodentia, rodentia, …
#> $ nome_familia        <fct> dasyproctidae, dasyproctidae, dasyproctidae, dasyp…
#> $ nome_genero         <fct> dasyprocta, dasyprocta, dasyprocta, dasyprocta, da…
#> $ nome_sp             <fct> dasyprocta_croconota, dasyprocta_croconota, dasypr…
#> $ nome_sp_abv         <fct> dasy_croc, dasy_croc, dasy_croc, dasy_croc, dasy_c…
#> $ validacao           <fct> especie, especie, especie, especie, especie, espec…
#> $ distancia           <dbl> 10, 0, 5, 12, 18, 15, 10, 10, 20, 20, 10, 0, 0, 10…
#> $ tamanho_grupo       <dbl> 1, 1, 1, 1, 1, 2, 1, 2, 1, 2, 1, 1, 1, 1, 2, 1, 2,…
#> $ velocidade_km_h     <dbl> 0.5555556, NA, NA, 0.7692308, NA, NA, NA, 0.625000…
#> $ tempo_censo         <drtn> 540 mins, 540 mins, 540 mins, 390 mins, 390 mins,…
#> $ numero_observadores <dbl> 1, 1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2,…

# carregar dados filtrados por Unidades de Conservação, espécie e nível taxonômico de validação
dados_filtrados_uc_sp_validacao <- filtrar_dados(
  nome_ucs = c("resex_tapajos_arapiuns", "resex_barreiro_das_antas"),
  nome_sps = c("dasyprocta_croconota", "guerlinguetus_aestuans"),
  validacao_obs = "especie"
)

glimpse(dados_filtrados_uc_sp_validacao)
#> Rows: 1,384
#> Columns: 24
#> $ nome_ea             <fct> ana, ana, ana, ana, ana, ana, ana, ana, ana, ana, …
#> $ n_visitas_repetidas <int> 67, 67, 67, 67, 67, 67, 67, 67, 67, 67, 67, 67, 67…
#> $ codigo_uc           <dbl> 259, 259, 259, 259, 259, 259, 259, 259, 259, 259, …
#> $ categoria_uc        <fct> resex, resex, resex, resex, resex, resex, resex, r…
#> $ nome_uc             <fct> resex_tapajos_arapiuns, resex_tapajos_arapiuns, re…
#> $ nome_uc_abv         <fct> dasy_croc, dasy_croc, guer_aest, dasy_croc, dasy_c…
#> $ numero_ea           <dbl> 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6,…
#> $ data_amostragem     <dttm> 2015-02-25, 2015-02-25, 2015-02-27, 2015-02-27, 2…
#> $ estacao             <fct> chuvosa, chuvosa, chuvosa, chuvosa, chuvosa, chuvo…
#> $ ano                 <dbl> 2015, 2015, 2015, 2015, 2015, 2015, 2015, 2015, 20…
#> $ esforco_dia         <dbl> 5000, 5000, 5000, 5000, 5000, 5000, 5000, 5000, 50…
#> $ esforco_total       <dbl> 335000, 335000, 335000, 335000, 335000, 335000, 33…
#> $ nome_classe         <fct> mammalia, mammalia, mammalia, mammalia, mammalia, …
#> $ nome_ordem          <fct> rodentia, rodentia, rodentia, rodentia, rodentia, …
#> $ nome_familia        <fct> dasyproctidae, dasyproctidae, sciuridae, dasyproct…
#> $ nome_genero         <fct> dasyprocta, dasyprocta, guerlinguetus, dasyprocta,…
#> $ nome_sp             <fct> dasyprocta_croconota, dasyprocta_croconota, guerli…
#> $ nome_sp_abv         <fct> dasy_croc, dasy_croc, guer_aest, dasy_croc, dasy_c…
#> $ validacao           <fct> especie, especie, especie, especie, especie, espec…
#> $ distancia           <dbl> 10, 0, 3, 5, 12, 18, 15, 10, 10, 20, 20, 10, 0, 0,…
#> $ tamanho_grupo       <dbl> 1, 1, 1, 1, 1, 1, 2, 1, 2, 1, 2, 1, 1, 1, 1, 2, 1,…
#> $ velocidade_km_h     <dbl> 0.5555556, NA, 0.5555556, NA, 0.7692308, NA, NA, N…
#> $ tempo_censo         <drtn> 540 mins, 540 mins, 540 mins, 540 mins, 390 mins,…
#> $ numero_observadores <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 2, 2, 2, 2,…