Seleciona, trasforma e renomea as colunas e observações nos dados brutos

A função gerar_dados_completos() recebe uma tibble gerada a partir da função carregar_dados_brutos_xlsx(), selecionando, transformando e renomeando suas colunas e observações.

Usage

gerar_dados_completos(dados)

Arguments

dados

recebe uma tibble contendo os dados brutos.

Value

Retorna um objeto do tipo tibble contendo uma seleção de colunas transformadas e renomeadas a partir dos dados brutos do Projeto Monitora Componente Florestal.

Details

A funcao gerar_dados_completos() seleciona, transforma e renomeia as colunas e observações dos dados brutos do Projeto Monitora Componente Florestal.

A função gera novas colunas contendo o número de vezes que cada estação amostral foi visitada, a categoria da UC, o nome abreviado das UC's, o nome abreviado das espécies, o esforço amostral total empregado em cada estação amostral, o número de observadores e o tempo total de cada censo. Quando ausentes, as distâncias de trilhas percorridas sao imputadas. Outra transfomração importante é a inlcusão de linhas para os dias em que houve amostragem e uma dada espécie não foi observada (ex. distância = NA)

Author

Vitor N. T. Borges-Junior Luciana A. Fusinatto

Examples

if (FALSE) gerar_dados_completos(dados = dados_brutos) # \dontrun{}