Estimativa de incerteza para modelos distance sampling por bootstrap

Conduz um bootstrap para modelos de distance sampling.

Usage

gerar_intervalos_confianca_bootstrap(
  dados,
  funcao_chave = c("hn", "hr", "unif"),
  termos_ajuste = NULL,
  truncamento = NULL,
  formula = ~1,
  funcao_sumario = c("abundancia", "densidade"),
  nboot = 100,
  nucleos = 1
)

Arguments

dados

recebe um conjunto de dados filtrados pela função filtrar_dados() e transformados pela função transformar_dados_formato_Distance().

funcao_chave

recebe uma dentre três funções utilizadas para modelar a detectabilidade da espécie: "hn" - Half-normal; "hr" - Hazard-rate; e "unif" - Uniform.

termos_ajuste

recebe um dentre n termos de ajuste: "cos" - cosseno; "herm" - Hermet polynomial; e "poly" - Polinomial simples. Quando NULL diferentes modelos são ajustados para função chave escolhida utilizando diferentes termos de ajuste. Veja Details.

truncamento

recebe uma lista contendo as distâncias de truncagem de forma numérica (ex. 20), ou como porcentagem (como um caracter, ex. 25%). Também pode ser alimentada no formato de lista, com os elementos left e right (ex. list(left = 1, right =20)) se a truncagem a esquerda for necessária. Por configuração, a distância máxima é utilizada como valor de truncagem a direita. Quando os dados estão categorizados, a truncagem a direita é o valor final da última coluna. O valor de truncagem a esquerda é zero, por configuração.

formula

a fórmula para o parâmetro de escala. Para análise CDS manter a configuração original ~ 1.

funcao_sumario

defini se o bootstrap retornará intervalos de confiança para estimativa de abunância ou de densidade.

nboot

número de bootstraps a serem conduzidos.

nucleos

número de núcleos do CPU usados para realizar a estimativa do intervalo de confiança.

Value

Retornar um vetor contendo o intervalo de confiança e um gáfico de histograma contendo a distribuição de frequências da estimativa de abundância ou de densidade e seus respectivos intervalos de confiança.

Details

Quando termo_ajuste = NULL modelos com diferente termos de ajuste serão ajustados aos dados, a depender da função chave utilizada. Para funcao_chave = "hn", são utilizados os termos de ajuste NULL (sem termos), "cos" e "herm"; para funcao_chave = "hr", são utilizados os termos de ajuste NULL (sem termos), "cos" e "poly"; e para funcao_chave = "unif", são utilizados so termos de ajuste "cos" e "poly".

Examples

if (FALSE) { # \dontrun{
gerar_intervalos_confianca_bootstrap()
} # }