Dados brutos de aves e mamiferos do Projeto Monitora Componente Florestal

Traz os dados de aves e mamiferos de medio e grande porte obtidos por amostragem por distancia em 40 Unidades de Conservacao do Brasil entre os anos de 2014 e 2022.

Usage

monitora_aves_masto_florestal

Format

'monitora_aves_masto_florestal'

Uma tibble que contem 26.420 linhas de 19 colunas:

nome_ea

nome da Estacao Amostral

n_visitas_repetidas

numero total de visitas realizadas em cada Estacao Amostral

codigo_uc

codigo da Unidade de Conservacao

categoria_uc

categoria da Unidade de Conservacao

nome_uc

nome da Unidade de Conservacao

nome_uc_abv

nome da Unidade de Conservacao abreviado

numero_ea

numero da Estacao Amostral dentro de uma Unidade de Conservacao

data_amostragem

ano, mes e dia em que a observacao foi realizada

estacao

estacao do ano em que a observacao foi realizada

ano

ano em que a observacao foi realizada

esforco_dia

distancia total percorrida em metros em um dia da amostragem na Estacao Amostral equivalente

esforco_total

distancia total percorrida em metros em todos os dias de amostragem na Estacao Amostral Equivalente

nome_classe

nome da classe ao qual o individuo observado pertence

nome_ordem

nome da ordem ao qual o individuo observado pertence

nome_familia

nome do família ao qual o individuo observado pertence

nome_genero

nome do gênero ao qual o individuo observado pertence

nome_sp

nome do nivel taxonomico ao qual o individuo observado pertence, gênero sp. ou espécie

nome_sp_abv

nome abreviado do nivel taxonomico do individuo observado, ex. se o individuo foi identificado ao nivel de familia, corresponde a familia, se foi identificado ao nivel de genero, corresponde ao genero, se foi identificado ao nivel de especie, corresponde a especie

validacao

nivel taxonomico de validacao do individuo observado

distancia

distancia perpendicular em relacao ao amostrador na qual o individuo foi observado

tamanho_grupo

numero de individuos registrados na mesma observacao

numero_observadores

numero de observadores que precorriam o transecto na ocasiao da observacao

velocidade_km_h

velocidade média do percurso do transecto

tempo_censo

tempo total de duracao do percurso de uma estacao amostral

Source

https://www.gov.br/icmbio/pt-br/assuntos/monitoramento/conteudo/dados