Transforma os dados para o formato para análise no pacote Distance

A função transformar_dados_formato_Distance() transforma a base de dados para o formato de análise do pacote Distance podendo retornar a base de dados com ou sem amostras repetidas na estações amostrais.

Usage

transformar_dados_formato_Distance(
  dados = monitora_aves_masto_florestal,
  ...,
  amostras_repetidas = TRUE
)

Arguments

dados

recebe a tibble gerada pela função filtrar_dados(). Por configuraçãoo, carrega a base de dados burtos de aves e médios e grandes mamíferos do Pojeto Monitora Componente Florestal monitora_aves_masto_florestal.

...

recebe a estratificação dos dados

amostras_repetidas

lógico, controla a saída dos dados transformados. Por configuração, retorna os dados com amostragens repetidas em cada estação amostral.

@details Quando o agrumento amostras_repetidas = FALSE será selecionado o dia em que foi registrado o maior número de observações em uma estação amostral.

Value

Retorna um objeto do tipo tibble contendo a base de dados configurada para ajustar os modelos do pacote Distance a partir da função [ajustar_modelo_distance().

Examples

# carregar pacote
library(dplyr)

# gerar os dados filtrados
dados_filtrados <- filtrar_dados(
  dados = monitora_aves_masto_florestal,
  nome_uc == "resex_tapajos_arapiuns",
  nome_sp == "dasyprocta_croconota",
  validacao_obs = "especie"
)

# gerar os dados transformados com repeticao
dados_distance_com_repeticao <- transformar_dados_formato_Distance(
  dados = dados_filtrados,
  Region.Label,
  Sample.Label,
  season,
  year
)

glimpse(dados_distance_com_repeticao)
#> Rows: 1,355
#> Columns: 14
#> $ Region.Label    <fct> resex_tapajos_arapiuns, resex_tapajos_arapiuns, resex_…
#> $ Area            <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ Sample.Label    <fct> boim, boim, boim, boim, boim, boim, boim, boim, boim, …
#> $ repeated_visits <int> 1, 12, 12, 12, 12, 12, 12, 12, 12, 12, 12, 12, 12, 29,…
#> $ Effort_day      <dbl> 5000, 5000, 5000, 5000, 5000, 5000, 5000, 5000, 5000, …
#> $ Effort          <dbl> 5000, 60000, 60000, 60000, 60000, 60000, 60000, 60000,…
#> $ sampling_day    <dttm> 2015-02-06, 2015-06-28, 2015-06-29, 2015-06-30, 2015-…
#> $ distance        <dbl> 10.00, 20.00, 28.00, 30.00, 4.00, 10.00, 20.00, 6.00, …
#> $ season          <fct> chuvosa, seca, seca, seca, seca, seca, seca, seca, sec…
#> $ year            <dbl> 2015, 2015, 2015, 2015, 2015, 2015, 2015, 2015, 2015, …
#> $ size            <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 2, 1, 1, 1, 1, 1, 1, 1, 1, 3, …
#> $ cense_time      <drtn> 525 mins, 600 mins, 600 mins, 645 mins, 215 mins, 215…
#> $ speed           <dbl> 0.5714286, 0.5000000, 0.5000000, 0.4651163, 1.3953488,…
#> $ object          <int> 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16,…

# gerar os dados transformados com repeticao
dados_distance_sem_repeticao <- transformar_dados_formato_Distance(
  dados = dados_filtrados,
  Region.Label,
  Sample.Label,
  season,
  year,
  amostras_repetidas = FALSE
)

glimpse(dados_distance_sem_repeticao)
#> Rows: 499
#> Columns: 14
#> $ Region.Label    <fct> resex_tapajos_arapiuns, resex_tapajos_arapiuns, resex_…
#> $ Area            <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ Sample.Label    <fct> boim, boim, boim, boim, boim, boim, boim, boim, boim, …
#> $ repeated_visits <int> 1, 12, 12, 12, 12, 29, 29, 11, 11, 11, 11, 13, 13, 13,…
#> $ Effort_day      <dbl> 5000, 5000, 5000, 5000, 5000, 5000, 5000, 5000, 5000, …
#> $ Effort          <dbl> 5000, 60000, 60000, 60000, 60000, 145000, 145000, 5500…
#> $ sampling_day    <dttm> 2015-02-06, 2015-10-23, 2015-10-23, 2015-10-23, 2015-…
#> $ distance        <dbl> 10.00, 10.00, 6.00, 30.00, 4.00, 30.00, 5.00, 8.34, 6.…
#> $ season          <fct> chuvosa, seca, seca, seca, seca, chuvosa, chuvosa, sec…
#> $ year            <dbl> 2015, 2015, 2015, 2015, 2015, 2016, 2016, 2016, 2016, …
#> $ size            <dbl> 1, 2, 1, 1, 1, 1, 3, 1, 1, 1, 2, 1, 1, 1, 2, 1, 1, 1, …
#> $ cense_time      <drtn> 525 mins, 78 mins, 78 mins, 78 mins, 78 mins, 228 min…
#> $ speed           <dbl> 0.5714286, 3.8461538, 3.8461538, 3.8461538, 3.8461538,…
#> $ object          <int> 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16,…