Número total de observações para cada espécie e em cada ano

A função contar_n_obs_sp_ano() gera uma tabela contendo o número total de observações para cada espécie e em cada ano.

Usage

contar_n_obs_sp_ano(dados = monitora_aves_masto_florestal)

Arguments

dados

recebe a tibble gerada pela função filtrar_dados(). Por configuração, carrega a base de dados burtos de aves e médios e grandes mamíferos do Pojeto Monitora Componente Florestal monitora_aves_masto_florestal.

Value

Retorna um objeto do tipo tibble contendo quatro colunas, ano que contem o ano em que a espécie foi amostrada, nome_sp contendo o nome da espécie, nome_sp_abv que contem o nome abreviado da espécie e n contendo o número de observações realizadas para espécie em cada ano.

Author

Vitor N. T. Borges-Junior Luciana A. Fusinatto

Examples

# número de observações para cada espécie por ano - dados brutos
contar_n_obs_sp_ano()
#> # A tibble: 940 × 4
#>      ano nome_sp                 nome_sp_abv     n
#>    <dbl> <fct>                   <fct>       <int>
#>  1  2014 alouatta_macconnelli    alou_macc       6
#>  2  2014 alouatta_nigerrima      alou_nige       4
#>  3  2014 alouatta_puruensis      alou_puru       8
#>  4  2014 aotus_nigriceps         aotu_nigr       7
#>  5  2014 ateles_chamek           atel_cham       3
#>  6  2014 ateles_marginatus       atel_marg       1
#>  7  2014 ateles_paniscus         atel_pani      28
#>  8  2014 brachyteles_arachnoides brac_arac       2
#>  9  2014 bradypus_variegatus     brad_vari       2
#> 10  2014 cabassous_unicinctus    caba_unic       2
#> # ℹ 930 more rows

# número de observações para cada espécie por ano - dados filtrados
dados_filtrados <- filtrar_dados(
  nome_ucs = "resex_tapajos_arapiuns"
)

contar_n_obs_sp_ano(dados_filtrados)
#> # A tibble: 263 × 4
#>      ano nome_sp                nome_sp_abv     n
#>    <dbl> <fct>                  <fct>       <int>
#>  1  2014 alouatta_nigerrima     alou_nige       4
#>  2  2014 bradypus_variegatus    brad_vari       2
#>  3  2014 callicebus_hoffmannsi  call_hoff       3
#>  4  2014 cebus_unicolor         cebu_unic       1
#>  5  2014 chiropotes_albinasus   chir_albi       2
#>  6  2014 crypturellus_sp        cryp_sp         4
#>  7  2014 dasyprocta_croconota   dasy_croc       5
#>  8  2014 eira_barbara           eira_barb       2
#>  9  2014 guerlinguetus_aestuans guer_aest       1
#> 10  2014 mico_humeralifer       mico_hume       4
#> # ℹ 253 more rows