Filtra os dados a partir do nível taxonomico de validação e demais colunas

A função filtrar_dados() filtra a base de dados brutos a partir da coluna validacao e colunas desejadas.

Usage

filtrar_dados(
         dados = monitora_aves_masto_florestal,
         ...,
         validacao_obs = c("ordem", "familia", "genero", "especie", "na")
       )

Arguments

dados

recebe uma tibble que contenha as colunas nome_uc, nome_sp e validacao. Por configuração, carrega a base de dados burto de aves e médios e grandes mamíferos do Pojeto Monitora Componente Florestal, monitora_aves_mamiferos_florestal.

...

recebe as colunas a sere filtradas ex. nome_uc, nome_sp, nome_genero.

validacao_obs

recebe um vetor do tipo caracter contendo o nome de um ou mais níveis taxonômicos de validação.

Value

Retorna um objeto do tipo tibble contendo os dados do Projeto Monitora Componente Florestal, filtrados a partir da(s) Unidade(s) de Conservação, da(s) espécie(s), ou dos nível(is) taxonômico(s) de validação.

Details

A filtragem da base de dados pode ser feita a partir do nome das Unidades de Conservação, a partir do nome das espécies, a partir do nível taxonômico de validação, ou a partir de suas combinações. Para maiores detalhes consulte os exemplos.

Author

Vitor N. T. Borges-Junior Luciana A. Fusinatto

Examples

# carregar pacote
library(dplyr)

# carregar dados filtrados por uma Unidade de Conservacão
dados_filtrados_uc1 <- filtrar_dados(
dados = monitora_aves_masto_florestal,
nome_uc == "resex_tapajos_arapiuns"
)

glimpse(dados_filtrados_uc1)
#> Rows: 3,582
#> Columns: 22
#> $ categoria_uc        <fct> resex, resex, resex, resex, resex, resex, resex, r…
#> $ nome_uc             <fct> resex_tapajos_arapiuns, resex_tapajos_arapiuns, re…
#> $ nome_uc_abv         <fct> rese_tapa, rese_tapa, rese_tapa, rese_tapa, rese_t…
#> $ codigo_uc           <dbl> 259, 259, 259, 259, 259, 259, 259, 259, 259, 259, …
#> $ nome_ea             <fct> boim, boim, boim, boim, boim, boim, boim, boim, bo…
#> $ numero_ea           <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,…
#> $ data_amostragem     <dttm> 2015-02-06, 2015-02-06, 2015-02-06, 2015-02-07, 2…
#> $ estacao             <fct> chuvosa, chuvosa, chuvosa, chuvosa, chuvosa, chuvo…
#> $ ano                 <dbl> 2015, 2015, 2015, 2015, 2015, 2015, 2015, 2015, 20…
#> $ esforco_dia         <dbl> 5000, 5000, 5000, 5000, 5000, 5000, 5000, 5000, 50…
#> $ nome_classe         <fct> mammalia, mammalia, mammalia, aves, mammalia, aves…
#> $ nome_ordem          <fct> artiodactyla, rodentia, primates, galliformes, art…
#> $ nome_familia        <fct> cervidae, dasyproctidae, cebidae, cracidae, tayass…
#> $ nome_genero         <fct> mazama, dasyprocta, sapajus, na, pecari, psophia, …
#> $ nome_sp             <fct> mazama_sp, dasyprocta_croconota, sapajus_apella, c…
#> $ nome_sp_abv         <fct> maza_sp, dasy_croc, sapa_apel, crac, peca_taja, ps…
#> $ validacao           <fct> genero, especie, especie, familia, especie, especi…
#> $ distancia           <dbl> 20, 10, 20, 15, 30, 20, 20, 18, 28, 10, 30, 20, 5,…
#> $ tamanho_grupo       <dbl> 1, 1, 5, 1, 1, 4, 1, 1, 1, 1, 1, 3, 2, 1, 1, 1, 1,…
#> $ velocidade_km_h     <dbl> 0.5714286, 0.5714286, 0.5714286, 0.7142857, 0.7142…
#> $ tempo_censo         <drtn> 525 mins, 525 mins, 525 mins, 420 mins, 420 mins,…
#> $ numero_observadores <dbl> 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2,…

# carregar dados filtrados por mais de uma Unidade de Conservação
dados_filtrados_uc2 <- filtrar_dados(
dados = monitora_aves_masto_florestal,
nome_uc %in% c("resex_tapajos_arapiuns", "resex_barreiro_das_antas")
)

glimpse(dados_filtrados_uc2)
#> Rows: 4,142
#> Columns: 22
#> $ categoria_uc        <fct> resex, resex, resex, resex, resex, resex, resex, r…
#> $ nome_uc             <fct> resex_barreiro_das_antas, resex_barreiro_das_antas…
#> $ nome_uc_abv         <fct> rese_barr_das_anta, rese_barr_das_anta, rese_barr_…
#> $ codigo_uc           <dbl> 221, 221, 221, 221, 221, 221, 221, 221, 221, 221, …
#> $ nome_ea             <fct> joao_dantas, joao_dantas, joao_dantas, joao_dantas…
#> $ numero_ea           <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,…
#> $ data_amostragem     <dttm> 2017-08-03, 2017-08-03, 2017-08-03, 2017-08-04, 2…
#> $ estacao             <fct> seca, seca, seca, seca, seca, seca, seca, seca, se…
#> $ ano                 <dbl> 2017, 2017, 2017, 2017, 2017, 2017, 2017, 2017, 20…
#> $ esforco_dia         <dbl> 5000, 5000, 5000, 5000, 5000, 5000, 5000, 5000, 50…
#> $ nome_classe         <fct> aves, mammalia, aves, aves, aves, mammalia, mammal…
#> $ nome_ordem          <fct> tinamiformes, primates, galliformes, galliformes, …
#> $ nome_familia        <fct> tinamidae, cebidae, cracidae, cracidae, tinamidae,…
#> $ nome_genero         <fct> tinamus, sapajus, penelope, penelope, tinamus, sap…
#> $ nome_sp             <fct> tinamus_sp, sapajus_apella, penelope_sp, penelope_…
#> $ nome_sp_abv         <fct> tina_sp, sapa_apel, pene_sp, pene_sp, tina_sp, sap…
#> $ validacao           <fct> genero, especie, genero, genero, genero, especie, …
#> $ distancia           <dbl> 0.00, 13.20, 2.35, 10.14, 6.00, 12.25, 15.13, 0.00…
#> $ tamanho_grupo       <dbl> 1, 3, 1, 2, 1, 4, 5, 2, 7, 1, 6, 2, 1, 4, 2, 3, 1,…
#> $ velocidade_km_h     <dbl> 1.0676157, 1.0676157, 1.0676157, 1.0169492, 1.0169…
#> $ tempo_censo         <drtn> 281 mins, 281 mins, 281 mins, 295 mins, 292 mins,…
#> $ numero_observadores <dbl> 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2,…

# carregar dados filtrados por mais de uma Unidade de Conservação e níveis taxonômicos de validação
dados_filtrados_uc_validacao1 <- filtrar_dados(
dados = monitora_aves_masto_florestal,
nome_uc == "resex_tapajos_arapiuns",
validacao_obs = c("genero", "especie")
)

glimpse(dados_filtrados_uc_validacao1)
#> Rows: 3,582
#> Columns: 22
#> $ categoria_uc        <fct> resex, resex, resex, resex, resex, resex, resex, r…
#> $ nome_uc             <fct> resex_tapajos_arapiuns, resex_tapajos_arapiuns, re…
#> $ nome_uc_abv         <fct> rese_tapa, rese_tapa, rese_tapa, rese_tapa, rese_t…
#> $ codigo_uc           <dbl> 259, 259, 259, 259, 259, 259, 259, 259, 259, 259, …
#> $ nome_ea             <fct> boim, boim, boim, boim, boim, boim, boim, boim, bo…
#> $ numero_ea           <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,…
#> $ data_amostragem     <dttm> 2015-02-06, 2015-02-06, 2015-02-06, 2015-02-07, 2…
#> $ estacao             <fct> chuvosa, chuvosa, chuvosa, chuvosa, chuvosa, chuvo…
#> $ ano                 <dbl> 2015, 2015, 2015, 2015, 2015, 2015, 2015, 2015, 20…
#> $ esforco_dia         <dbl> 5000, 5000, 5000, 5000, 5000, 5000, 5000, 5000, 50…
#> $ nome_classe         <fct> mammalia, mammalia, mammalia, aves, mammalia, aves…
#> $ nome_ordem          <fct> artiodactyla, rodentia, primates, galliformes, art…
#> $ nome_familia        <fct> cervidae, dasyproctidae, cebidae, cracidae, tayass…
#> $ nome_genero         <fct> mazama, dasyprocta, sapajus, na, pecari, psophia, …
#> $ nome_sp             <fct> mazama_sp, dasyprocta_croconota, sapajus_apella, c…
#> $ nome_sp_abv         <fct> maza_sp, dasy_croc, sapa_apel, crac, peca_taja, ps…
#> $ validacao           <fct> genero, especie, especie, familia, especie, especi…
#> $ distancia           <dbl> 20, 10, 20, 15, 30, 20, 20, 18, 28, 10, 30, 20, 5,…
#> $ tamanho_grupo       <dbl> 1, 1, 5, 1, 1, 4, 1, 1, 1, 1, 1, 3, 2, 1, 1, 1, 1,…
#> $ velocidade_km_h     <dbl> 0.5714286, 0.5714286, 0.5714286, 0.7142857, 0.7142…
#> $ tempo_censo         <drtn> 525 mins, 525 mins, 525 mins, 420 mins, 420 mins,…
#> $ numero_observadores <dbl> 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2,…

# carregar dados filtrados por uma espécie
dados_filtrados_sp1 <- filtrar_dados(
dados = monitora_aves_masto_florestal,
nome_sp == "dasyprocta_croconota"
)

glimpse(dados_filtrados_sp1)
#> Rows: 2,204
#> Columns: 22
#> $ categoria_uc        <fct> esec, esec, esec, esec, esec, esec, esec, esec, es…
#> $ nome_uc             <fct> esec_da_terra_do_meio, esec_da_terra_do_meio, esec…
#> $ nome_uc_abv         <fct> esec_da_terr_do_meio, esec_da_terr_do_meio, esec_d…
#> $ codigo_uc           <dbl> 47, 47, 47, 47, 47, 47, 47, 47, 47, 47, 47, 47, 47…
#> $ nome_ea             <fct> trilha_da_fazenda, trilha_da_fazenda, trilha_da_fa…
#> $ numero_ea           <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,…
#> $ data_amostragem     <dttm> 2014-10-22, 2014-10-22, 2014-10-23, 2014-10-23, 2…
#> $ estacao             <fct> seca, seca, seca, seca, seca, seca, seca, seca, se…
#> $ ano                 <dbl> 2014, 2014, 2014, 2014, 2014, 2014, 2014, 2014, 20…
#> $ esforco_dia         <dbl> 4650, 4650, 4650, 4650, 4650, 4650, 4650, 4650, 46…
#> $ nome_classe         <fct> mammalia, mammalia, mammalia, mammalia, mammalia, …
#> $ nome_ordem          <fct> rodentia, rodentia, rodentia, rodentia, rodentia, …
#> $ nome_familia        <fct> dasyproctidae, dasyproctidae, dasyproctidae, dasyp…
#> $ nome_genero         <fct> dasyprocta, dasyprocta, dasyprocta, dasyprocta, da…
#> $ nome_sp             <fct> dasyprocta_croconota, dasyprocta_croconota, dasypr…
#> $ nome_sp_abv         <fct> dasy_croc, dasy_croc, dasy_croc, dasy_croc, dasy_c…
#> $ validacao           <fct> especie, especie, especie, especie, especie, espec…
#> $ distancia           <dbl> 7.75, 1.70, 13.71, 11.90, 1.27, 6.92, 9.20, 9.35, …
#> $ tamanho_grupo       <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 2, 1, 1, 1,…
#> $ velocidade_km_h     <dbl> 0.9964286, 0.9964286, 1.0035971, 1.0035971, 0.9964…
#> $ tempo_censo         <drtn> 280 mins, 280 mins, 278 mins, 278 mins, 280 mins,…
#> $ numero_observadores <dbl> 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2,…

# carregar dados filtrados por mais de uma especie
dados_filtrados_sp2 <- filtrar_dados(
dados = monitora_aves_masto_florestal,
nome_sp %in% c("dasyprocta_croconota", "dasyprocta_iacki")
)

glimpse(dados_filtrados_sp2)
#> Rows: 2,210
#> Columns: 22
#> $ categoria_uc        <fct> esec, esec, esec, esec, esec, esec, esec, esec, es…
#> $ nome_uc             <fct> esec_da_terra_do_meio, esec_da_terra_do_meio, esec…
#> $ nome_uc_abv         <fct> esec_da_terr_do_meio, esec_da_terr_do_meio, esec_d…
#> $ codigo_uc           <dbl> 47, 47, 47, 47, 47, 47, 47, 47, 47, 47, 47, 47, 47…
#> $ nome_ea             <fct> trilha_da_fazenda, trilha_da_fazenda, trilha_da_fa…
#> $ numero_ea           <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,…
#> $ data_amostragem     <dttm> 2014-10-22, 2014-10-22, 2014-10-23, 2014-10-23, 2…
#> $ estacao             <fct> seca, seca, seca, seca, seca, seca, seca, seca, se…
#> $ ano                 <dbl> 2014, 2014, 2014, 2014, 2014, 2014, 2014, 2014, 20…
#> $ esforco_dia         <dbl> 4650, 4650, 4650, 4650, 4650, 4650, 4650, 4650, 46…
#> $ nome_classe         <fct> mammalia, mammalia, mammalia, mammalia, mammalia, …
#> $ nome_ordem          <fct> rodentia, rodentia, rodentia, rodentia, rodentia, …
#> $ nome_familia        <fct> dasyproctidae, dasyproctidae, dasyproctidae, dasyp…
#> $ nome_genero         <fct> dasyprocta, dasyprocta, dasyprocta, dasyprocta, da…
#> $ nome_sp             <fct> dasyprocta_croconota, dasyprocta_croconota, dasypr…
#> $ nome_sp_abv         <fct> dasy_croc, dasy_croc, dasy_croc, dasy_croc, dasy_c…
#> $ validacao           <fct> especie, especie, especie, especie, especie, espec…
#> $ distancia           <dbl> 7.75, 1.70, 13.71, 11.90, 1.27, 6.92, 9.20, 9.35, …
#> $ tamanho_grupo       <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 2, 1, 1, 1,…
#> $ velocidade_km_h     <dbl> 0.9964286, 0.9964286, 1.0035971, 1.0035971, 0.9964…
#> $ tempo_censo         <drtn> 280 mins, 280 mins, 278 mins, 278 mins, 280 mins,…
#> $ numero_observadores <dbl> 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2,…

# carregar dados filtrados por unidade de conservacao e especie
dados_filtrados_uc_sp <- filtrar_dados(
dados = monitora_aves_masto_florestal,
nome_uc == "resex_tapajos_arapiuns",
nome_sp == c("dasyprocta_croconota")
)

glimpse(dados_filtrados_uc_sp)
#> Rows: 1,355
#> Columns: 22
#> $ categoria_uc        <fct> resex, resex, resex, resex, resex, resex, resex, r…
#> $ nome_uc             <fct> resex_tapajos_arapiuns, resex_tapajos_arapiuns, re…
#> $ nome_uc_abv         <fct> rese_tapa, rese_tapa, rese_tapa, rese_tapa, rese_t…
#> $ codigo_uc           <dbl> 259, 259, 259, 259, 259, 259, 259, 259, 259, 259, …
#> $ nome_ea             <fct> boim, boim, boim, boim, boim, boim, boim, boim, bo…
#> $ numero_ea           <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,…
#> $ data_amostragem     <dttm> 2015-02-06, 2015-06-28, 2015-06-29, 2015-06-30, 2…
#> $ estacao             <fct> chuvosa, seca, seca, seca, seca, seca, seca, seca,…
#> $ ano                 <dbl> 2015, 2015, 2015, 2015, 2015, 2015, 2015, 2015, 20…
#> $ esforco_dia         <dbl> 5000, 5000, 5000, 5000, 5000, 5000, 5000, 5000, 50…
#> $ nome_classe         <fct> mammalia, mammalia, mammalia, mammalia, mammalia, …
#> $ nome_ordem          <fct> rodentia, rodentia, rodentia, rodentia, rodentia, …
#> $ nome_familia        <fct> dasyproctidae, dasyproctidae, dasyproctidae, dasyp…
#> $ nome_genero         <fct> dasyprocta, dasyprocta, dasyprocta, dasyprocta, da…
#> $ nome_sp             <fct> dasyprocta_croconota, dasyprocta_croconota, dasypr…
#> $ nome_sp_abv         <fct> dasy_croc, dasy_croc, dasy_croc, dasy_croc, dasy_c…
#> $ validacao           <fct> especie, especie, especie, especie, especie, espec…
#> $ distancia           <dbl> 10.00, 20.00, 28.00, 30.00, 4.00, 10.00, 20.00, 6.…
#> $ tamanho_grupo       <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 2, 1, 1, 1, 1, 1, 1, 1, 1,…
#> $ velocidade_km_h     <dbl> 0.5714286, 0.5000000, 0.5000000, 0.4651163, 1.3953…
#> $ tempo_censo         <drtn> 525 mins, 600 mins, 600 mins, 645 mins, 215 mins,…
#> $ numero_observadores <dbl> 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2,…

# carregar dados filtrados por Unidades de Conservação, espécie e nível taxonômico de validação
dados_filtrados_uc_sp_validacao <- filtrar_dados(
dados = monitora_aves_masto_florestal,
  nome_uc %in% c("resex_tapajos_arapiuns", "resex_barreiro_das_antas"),
  nome_sp %in% c("dasyprocta_croconota", "guerlinguetus_aestuans"),
  validacao_obs = "especie"
)

glimpse(dados_filtrados_uc_sp_validacao)
#> Rows: 1,387
#> Columns: 22
#> $ categoria_uc        <fct> resex, resex, resex, resex, resex, resex, resex, r…
#> $ nome_uc             <fct> resex_barreiro_das_antas, resex_barreiro_das_antas…
#> $ nome_uc_abv         <fct> rese_barr_das_anta, rese_barr_das_anta, rese_barr_…
#> $ codigo_uc           <dbl> 221, 221, 221, 221, 221, 221, 221, 221, 259, 259, …
#> $ nome_ea             <fct> joao_dantas, joao_dantas, joao_dantas, igarape_azu…
#> $ numero_ea           <dbl> 1, 1, 1, 2, 2, 2, 3, 3, 1, 1, 1, 1, 1, 1, 1, 1, 1,…
#> $ data_amostragem     <dttm> 2017-08-08, 2018-04-17, 2020-08-14, 2018-02-23, 2…
#> $ estacao             <fct> seca, chuvosa, seca, chuvosa, seca, chuvosa, seca,…
#> $ ano                 <dbl> 2017, 2018, 2020, 2018, 2020, 2021, 2018, 2021, 20…
#> $ esforco_dia         <dbl> 5000, 5000, 5000, 5000, 5000, 5000, 5000, 5000, 50…
#> $ nome_classe         <fct> mammalia, mammalia, mammalia, mammalia, mammalia, …
#> $ nome_ordem          <fct> rodentia, rodentia, rodentia, rodentia, rodentia, …
#> $ nome_familia        <fct> sciuridae, sciuridae, sciuridae, sciuridae, sciuri…
#> $ nome_genero         <fct> guerlinguetus, guerlinguetus, guerlinguetus, guerl…
#> $ nome_sp             <fct> guerlinguetus_aestuans, guerlinguetus_aestuans, gu…
#> $ nome_sp_abv         <fct> guer_aest, guer_aest, guer_aest, guer_aest, guer_a…
#> $ validacao           <fct> especie, especie, especie, especie, especie, espec…
#> $ distancia           <dbl> 3.30, 5.00, 0.00, 3.00, 3.11, 5.19, 3.00, 4.07, 10…
#> $ tamanho_grupo       <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 2,…
#> $ velocidade_km_h     <dbl> 1.0489510, 1.0380623, 1.3157895, 1.0169492, 1.3043…
#> $ tempo_censo         <drtn> 286 mins, 289 mins, 228 mins, 295 mins, 230 mins,…
#> $ numero_observadores <dbl> 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2,…