Gera uma tabela dinâmica

A função gerar_tabdin() gera uma tabela dinâmica a partir de uma base de dados de entrada.

Usage

gerar_tabdin(dados = monitora_aves_masto_florestal, n_linhas = 1:1000)

Arguments

dados

recebe a tibble gerada pela função filtrar_dados(). Por configuração, carrega a base de dados burto de aves e médios e grandes mamíferos do Pojeto Monitora Componente Florestal, monitora_aves_mamiferos_florestal.

n_linhas

fornece o intervalo de linhas a ser gerado. Por configuração, é gerado o intervalo entre as linhas 1 e 1000.

Value

Retorna um objeto do tipo ...

Details

A velociadade de geração da tabela dinâmica e o tamanho do intervalo de linhas que pode ser gerado vairam de acordo com a configuração da máquina que executa a função.

Author

Vitor N. T. Borges-Junior Luciana A. Fusinatto

Examples

# gerar tabela dinâmica dos dados brutos
gerar_tabdin()

# gerar tabela dinâmica a partir dos dados filtrados
dados_filtrados <- filtrar_dados(
dados = monitora_aves_masto_florestal,
nome_sp == c("dasyprocta_croconota", "dasyprocta_iacki")
)
#> Warning: There were 2 warnings in `dplyr::filter()`.
#> The first warning was:
#> ℹ In argument: `nome_sp == c("dasyprocta_croconota", "dasyprocta_iacki")`.
#> Caused by warning in `==.default`:
#> ! longer object length is not a multiple of shorter object length
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 1 remaining warning.
#> Warning: There were 2 warnings in `dplyr::filter()`.
#> The first warning was:
#> ℹ In argument: `nome_sp == c("dasyprocta_croconota", "dasyprocta_iacki")`.
#> Caused by warning in `==.default`:
#> ! longer object length is not a multiple of shorter object length
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 1 remaining warning.
gerar_tabdin(dados_filtrados)